+++ package.json
@@ -2,24 +2,24 @@
   "name": "linqts",
   "version": "0.0.0-semantic-release",
   "description": "LinQ + TypeScript",
-  "main": "dist/linq.js",
-  "es2015": "linq.ts",
-  "umd:main": "dist/linq.umd.js",
-  "typescript": {
-    "definition": "dist/linq.d.ts"
-  },
+  "es2015": "index.ts",
+  "main": "dist/linqts.js",
+  "module": "dist/linqts.es.js",
+  "types": "dist/linqts.d.ts",
+  "umd:main": "dist/linqts.umd.js",
   "scripts": {
-    "build": "tsc linq.ts -t es3 -m commonjs -d --outDir dist &&  rollup -c rollup.config.ts",
-    "check-coverage": "nyc check-coverage --statements 100 --branches 90 --functions 98 --lines 100",
+    "build": "microbundle -o dist",
+    "check-coverage": "nyc check-coverage --statements 100 --branches 97 --functions 97 --lines 100",
     "commit": "git-cz",
-    "compile": "tsc",
+    "compile": "tsc index.ts -t es6 -m es6",
     "cover": "nyc --reporter=lcov npm t",
-    "docs": "typedoc --out ../docs/ linq.ts -m commonjs -t ES6",
+    "dev": "microbundle watch",
+    "docs": "typedoc --out ../docs/ index.ts -m commonjs -t ES6",
     "nyan": "ava --tap | tap-nyan",
     "prebuild": "rimraf dist",
     "precompile": "npm run pretty",
     "pretest": "npm run compile",
-    "pretty": "prettier linq.ts test/**/*.ts --write --no-semi --single-quote",
+    "pretty": "prettier index.ts test/**/*.ts --write --no-semi --single-quote",
     "report-coverage": "nyc report --reporter=text-lcov | coveralls",
     "semantic-release": "semantic-release pre && npm publish && semantic-release post",
     "start": "tsc -w",
@@ -45,7 +45,6 @@
     "url": "https://github.com/kutyel"
   },
   "license": "MIT",
-  "typings": "dist/linq.d.ts",
   "bugs": {
     "url": "https://github.com/kutyel/linq.ts/issues"
   },
@@ -57,13 +56,10 @@
     "coveralls": "^3.0.0",
     "cz-conventional-changelog": "^2.0.0",
     "ghooks": "^2.0.0",
+    "microbundle": "^0.4.0",
     "nyc": "^11.0.2",
     "prettier": "^1.5.3",
     "rimraf": "^2.6.1",
-    "rollup": "^0.55.0",
-    "rollup-plugin-commonjs": "^8.0.2",
-    "rollup-plugin-node-resolve": "^3.0.0",
-    "rollup-plugin-sourcemaps": "^0.4.2",
     "semantic-release": "^12.2.2",
     "ts-node": "^4.0.0",
     "tslint": "^5.0.0",
