+++ tsconfig.json
@@ -1,10 +1,8 @@
 {
-    "compilerOptions": {
-        "skipLibCheck": true,
-        "module": "commonjs",
-        "target": "es2015"
-    },
-    "exclude": [
-        "node_modules"
-    ]
+  "compilerOptions": {
+    "module": "commonjs",
+    "declaration": true,
+    "target": "es5"
+  },
+  "exclude": ["node_modules"]
 }
